Fix Rubocop Style/SoleNestedConditional

+++ hooks/pre/20-check-hammer-credentials.rb
@@ -18,13 +18,11 @@ def password_set?(path)
 end
 
 new_config_file = File.join(NEW_HAMMER_CONFIG_PATH, NEW_HAMMER_CONFIG_FILE)
-unless File.exist?(new_config_file)
-  # if there is foreman password or non-admin user set in any of the legacy configs
-  # create empty hammer foreman config to prevent installer from creating new one
-  if POSSIBLE_RECENT_CONFIG_PATHS.any? { |path| password_set?(path) }
+# if there is foreman password or non-admin user set in any of the legacy configs
+# create empty hammer foreman config to prevent installer from creating new one
+if !File.exist?(new_config_file) && POSSIBLE_RECENT_CONFIG_PATHS.any? { |path| password_set?(path) }
     FileUtils.mkdir_p(NEW_HAMMER_CONFIG_PATH)
     File.open(new_config_file, "w+") do |file|
       file.write("---\n:foreman: {}\n")
     end
   end
-end
